requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
discord==1.0.1
discord.py==1.4.1
dnspython==2.0.0
ffmpeg-python==0.2.0
future==0.18.2
google-api-core==1.22.1
google-api-python-client==1.10.1
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
multidict==4.7.6
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.0
PyNaCl==1.4.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
wavelink==0.9.4
yarl==1.5.1
